requirements.txt
rlgym[rl-rlviser]==2.0.1
torch>=2.0.0
numpy>=1.24.0
tensorboard>=2.13.0
wandb>=0.15.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
pyyaml>=6.0
tqdm>=4.65.0
